# internal/serve/static/openapi.yaml
openapi: "3.1.1"
info:
  title: redgiant
  version: "0.0.1"

paths:
  /health:
    get:
      responses:
        "200":
          description: Successful Response

  /api/about:
    get:
      tags: ["API"]
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/About"
  /api/state:
    get:
      tags: ["API"]
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/State"
  /api/devices:
    get:
      tags: ["API"]
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
  /api/data/{deviceID}/real:
    get:
      tags: ["API"]
      parameters:
        - in: path
          name: deviceID
          schema:
            type: integer
          required: true
        - in: query
          name: lang
          schema:
            type: string
            enum:
              - ch_CN
              - en_US
              - de_DE
              - nl_NL
              - pl_PL
        - in: query
          name: service
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RealMeasurement"
  /api/data/{deviceID}/direct:
    get:
      tags: ["API"]
      parameters:
        - in: path
          name: deviceID
          schema:
            type: integer
          required: true
        - in: query
          name: lang
          schema:
            type: string
            enum:
              - ch_CN
              - en_US
              - de_DE
              - nl_NL
              - pl_PL
        - in: query
          name: service
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DirectMeasurement"

components:
  schemas:
    About:
      properties:
        serialNumber:
          type: string
        version:
          type: string
        softwareVersion:
          type: string
        buildVersion:
          type: string
    State:
      properties:
        totalFaults:
          type: integer
        totalAlarms:
          type: integer
        wirelessConnection:
          type: boolean
        wifiConnection:
          type: boolean
        ethernet1Connection:
          type: boolean
        ethernet2Connection:
          type: boolean
        cloudConnection:
          type: boolean
    Device:
      properties:
        id:
          type: integer
        code:
          type: integer
        type:
          type: integer
        protocol:
          type: integer
        serialNumber:
          type: string
        name:
          type: string
        model:
          type: string
        special:
          type: string
        invType:
          type: integer
        physicalAddress:
          type: integer
        logicalAddress:
          type: integer
        linkStatus:
          type: integer
        initStatus:
          type: integer
    RealMeasurement:
      properties:
        i18nCode:
          type: string
        name:
          type: string
        value:
          type: string
        unit:
          type: string
    DirectMeasurement:
      properties:
        i18nCode:
          type: string
        name:
          type: string
        voltage:
          type: number
        voltageUnit:
          type: string
        current:
          type: number
        currentUnit:
          type: string